+++ packages/passkeys/passkeys/lib/relying_party_server/corbado/types/authentication.g.dart
@@ -86,6 +86,7 @@ CorbadoAuthenticationComplete _$CorbadoAuthenticationCompleteFromJson(
       clientDataJSON: json['clientDataJSON'] as String,
       authenticatorData: json['authenticatorData'] as String,
       signature: json['signature'] as String,
+      userHandle: json['userHandle'] as String,
     );
 
 Map<String, dynamic> _$CorbadoAuthenticationCompleteToJson(
@@ -94,4 +95,5 @@ Map<String, dynamic> _$CorbadoAuthenticationCompleteToJson(
       'clientDataJSON': instance.clientDataJSON,
       'authenticatorData': instance.authenticatorData,
       'signature': instance.signature,
+      'userHandle': instance.userHandle,
     };