+++ packages/passkeys/passkeys/lib/relying_party_server/types/authentication.dart
@@ -55,6 +55,7 @@ class AuthenticationCompleteRequest {
     required this.clientDataJSON,
     required this.authenticatorData,
     required this.signature,
+    required this.userHandle,
   });
 
   /// The id
@@ -71,4 +72,6 @@ class AuthenticationCompleteRequest {
 
   /// The signature
   final String signature;
+
+  final String userHandle;
 }