+++ packages/passkeys/passkeys_android/lib/passkeys_android.dart
@@ -52,6 +52,7 @@ class PasskeysAndroid extends PasskeysPlatform {
       clientDataJSON: r.clientDataJSON,
       authenticatorData: r.authenticatorData,
       signature: r.signature,
+      userHandle: r.userHandle
     );
   }