+++ packages/passkeys/passkeys_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+
+* Added userHandle to AuthenticateResponseType.
+
 ## 1.0.0
 
 * Better comments. Stable release.